# perf(is): check for single datatype when possible