Slightly lowers the ops/sec for `undefined` and `null` in favor of increasing them for everything else. `undefined` and `null` are expected to decrease because previously there were little-to-no operations before their checks; however, it might not matter in the long term because it's likely that people will much more frequently check against almost any other `DataType`.

Average increase overall: 10%
Average Increase for all except `undefined` and `null`: 16%

| test case         | before (ops/sec) | after (ops/sec) | difference |
|-------------------|------------------|-----------------|------------|
| undefined-valid   | 27,038,957       | 22,871,528      | -15%       |
| undefined-invalid | 27,368,894       | 23,254,362      | -15%       |
| null-valid        | 23,352,353       | 21,140,243      | -9%        |
| null-invalid      | 23,146,336       | 20,407,678      | -12%       |
| boolean-valid     | 12,349,800       | 15,247,814      | 23%        |
| boolean-invalid   | 12,541,613       | 15,214,705      | 21%        |
| number-valid      | 9,100,878        | 9,776,449       | 7%         |
| number-invalid    | 12,997,872       | 15,243,125      | 17%        |
| integer-valid     | 6,800,126        | 8,075,495       | 19%        |
| integer-invalid   | 8,756,741        | 9,561,322       | 9%         |
| natural-valid     | 6,815,298        | 7,726,053       | 13%        |
| natural-invalid   | 8,681,553        | 10,105,690      | 16%        |
| string-valid      | 5,130,817        | 5,234,438       | 2%         |
| string-invalid    | 13,243,061       | 15,311,125      | 16%        |
| function-valid    | 13,123,377       | 14,964,545      | 14%        |
| function-invalid  | 13,097,122       | 15,255,205      | 16%        |
| object-valid      | 12,507,796       | 14,559,489      | 16%        |
| object-invalid    | 12,050,531       | 15,663,236      | 30%        |
| array-valid       | 8,306,534        | 9,062,618       | 9%         |
| array-invalid     | 14,402,991       | 17,273,208      | 20%        |